requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
distlib==0.4.0
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.119.0
filelock==3.20.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
platformdirs==4.5.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
virtualenv==20.35.3
